google-cloud-adk/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
google-genai>=0.3.0
